test3: everything fixed

+++ lpp.py
@@ -80,14 +80,16 @@ def request(url):
             return requests.get("http://" + url)
         except requests.exceptions.ConnectionError:
             print("Error: Can't connect to website!")
-            os.system('rm dir_result.txt')
-            sys.exit(0)
+            if os.path.exists('dir_result.txt'):
+                os.system('rm dir_result.txt')
+                sys.exit(0)
+            else:
+                sys.exit(0)
 
 def test3():
     os.system('cls' if os.name == 'nt' else 'clear')
-    host=input("Enter IP | DNS : ")
-    #if (scan[host].state != "up"):
-     #   print("Your host is down")
+    host=input("Enter IP | DNS <<|http://| not needed>>: ")
+    check_dns=request(host)
     print("######   If you want to add your own file, put it in the folder [dir_list] ######")
     os.system('cd dir_list && ls | tr " " "\n" | tr ".txt" " "')
     dir_file =input("Choose a fuzz's file between this ones : ")
@@ -96,6 +98,11 @@ def test3():
     os.system('touch dir_result.txt')
     #try to found a way to display the number of tests like if we have 100 words in the wordlist, it display X/100
     f = open('dir_result.txt', 'w')
+    while (dir_file != os.path.exists("dir_list/"+dir_file)): #Infinite loop. Need to fix this
+        print("The file entered do not exist!!")
+        print("Try again... :")
+        os.system('cd dir_list && ls | tr " " "\n" | tr ".txt" " "')
+        dir_file =input("Choose a fuzz's file between this ones : ")
     if os.path.exists("dir_list/"+dir_file):
         file = open("dir_list/"+dir_file)
         for lines in file:
@@ -105,10 +112,12 @@ def test3():
             if response:
                 print("==============================================================================")
                 print("A directory has been found here: "+url)
-                f.write(str(url)) # DOESNT WORK!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-    print("====================================================================================")
-    file.close()
-    f.close()
+                f.write(str(url+"\n"))
+        print("====================================================================================")
+        file.close()
+        f.close()
+    else:
+        print("/!\ The Website Is Not Up /!\ ")
 
 
 os.system('cls' if os.name == 'nt' else 'clear')
